[feat] 당첨  번호 입력 받기 기능
