+++ src/App.js
@@ -1,5 +1,13 @@
+import { print } from "./view/print";
+import { readLineAsync } from "./view/readLineAsync";
+import { MESSAGE } from "./constant/MESSAGE";
+import Lotto from "./view/Lotto"
+
 class App {
-  async play() {}
+
+  async play() {
+  }
+  
 }
 
 export default App;
