+++ src/constant/ERROR_MESSAGE.js
@@ -1,3 +1,4 @@
 export const ERROR_MESSAGE = Object.freeze({
-	
+	LOTTO_LENGTH_ERROR: '[ERROR] 로또 번호는 6개여야 합니다.',
+	LOTTO_DUPLICATE_ERROR: '[ERROR] 로또 번호는 중복이 없어야 합니다.'
 })
