+++ src/constant/MESSAGE.js
@@ -1,3 +1,7 @@
 export const MESSAGE = Object.freeze({
-	
+	INPUT_AMOUNT: '구입금액을 입력해 주세요.',
+	TOTAL_PURCHASE: '개를 구매했습니다.',
+	INPUT_WINNING_NUMBERS: '당첨 번호를 입력해 주세요.',
+	INPUT_BONUS_NUMBER: '보너스 번호를 입력해 주세요.',
+	WINNING_STATICS: '당첨 통계',
 })
